Update README.md
This is our university 2nd Year First project.
About our Website is below ----- 

The department of Agriculture in Sri Lanka in company with Keels are planning to launch a website with the aim of developing a virtual market which facilitate the farmers to sell their harvest and then the Keels staff able to buy those. Farmers are allowed to register to the system by using their NIC numbers and details related to crop. Sign up facilities are available freely for farmers. 
The registered farmers are given the facility to input data such as images. After the harvest is picked, the farmers can upload the report to the website about their harvest along with some photos. They can delete or update those reports. The uploaded reports can viewed by the DoA and the Keels staff. They can validate the reports with the help of the photos that are uploaded and they can take required steps. And also they can give the approval for the reports of the farmers using the evidence that has been evaluated. Then they send requests to the reported locations which are shown on the map. The Keels staff can communicate with the farmers through direct messages and get the decision whether they want to buy the products or ignore them. They can set a flag to evaluate the quality of the product (red flag for terrible quality). Various flagged reports are shown on a map along with various icons. By clicking that icons they can view more information.
 DoA staff can view various graphs such as, analysis of products, locations where most farmers are live in, transactions which were successful, products which were wasted. General public can view the reports of the farmers on a map and graphs, but they can’t see the details of the farmers.